Second round update belindas-closet ReadMe
This is a little more polished and goes a little further in the direction I am hoping to take this document. Most of the ToDos are in reference to links I need to round up. Some of those links will come from the PO progress videos. I am looking for early comment by the TProjs to see if there are any holes. @taylorpapke @brinkbrink 

From @taylorpapke and @Robel-003, I am looking for the corresponding video links for each section. 

Thanks!